requirements-docker.txt
# Gemini Live Bot - Docker Requirements
# Minimal production dependencies

# Core framework with all necessary extras
pipecat-ai[silero,google,webrtc]==0.0.98

# Google Gemini API
google-generativeai==0.8.6

# Audio processing
numpy==2.2.6
scipy==1.17.0

# Environment management
python-dotenv==1.2.1

# Logging
loguru==0.7.3

# WebRTC dependencies (explicitly listed for Docker)
opencv-python==4.12.0.88
aiortc==1.14.0
aiohttp==3.13.3

# FastAPI for web server
fastapi==0.121.3
uvicorn[standard]==0.40.0

# Additional required dependencies
protobuf==5.29.5
pydantic==2.12.5
